third_party/WebKit/LayoutTests/platform/android/svg/animations/animate-elem-08-t-drt-expected.txt

SVG 1.1 dynamic animation tests


A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getTransformToElement(rootSVGElement, path1).a is 0
PASS getTransformToElement(rootSVGElement, path1).b is 1
PASS getTransformToElement(rootSVGElement, path1).c is -1
PASS getTransformToElement(rootSVGElement, path1).d is 0
PASS getTransformToElement(rootSVGElement, path1).e is 224.9
FAIL getTransformToElement(rootSVGElement, path1).f should be close to -25.9. Was -25.000000000000014.
PASS getTransformToElement(rootSVGElement, path2).a is 0
PASS getTransformToElement(rootSVGElement, path2).b is -1
PASS getTransformToElement(rootSVGElement, path2).c is 1
PASS getTransformToElement(rootSVGElement, path2).d is 0
FAIL getTransformToElement(rootSVGElement, path2).e should be close to -223.9. Was -225.
FAIL getTransformToElement(rootSVGElement, path2).f should be close to 275.9. Was 275.
PASS getTransformToElement(rootSVGElement, path1).a is 1
PASS getTransformToElement(rootSVGElement, path1).b is 0.1
PASS getTransformToElement(rootSVGElement, path1).c is -0.1
PASS getTransformToElement(rootSVGElement, path1).d is 1
FAIL getTransformToElement(rootSVGElement, path1).e should be close to -70.3. Was -70.87814895357428.
FAIL getTransformToElement(rootSVGElement, path1).f should be close to -182.8. Was -182.60109333201257.
PASS getTransformToElement(rootSVGElement, path2).a is -1
PASS getTransformToElement(rootSVGElement, path2).b is -0.1
PASS getTransformToElement(rootSVGElement, path2).c is 0.1
PASS getTransformToElement(rootSVGElement, path2).d is -1
FAIL getTransformToElement(rootSVGElement, path2).e should be close to 318.6. Was 319.29822182153924.
FAIL getTransformToElement(rootSVGElement, path2).f should be close to 211.6. Was 210.66313575977358.
PASS getTransformToElement(rootSVGElement, path1).a is 0.7
PASS getTransformToElement(rootSVGElement, path1).b is -0.7
PASS getTransformToElement(rootSVGElement, path1).c is 0.7
PASS getTransformToElement(rootSVGElement, path1).d is 0.7
PASS getTransformToElement(rootSVGElement, path1).e is -265.1
FAIL getTransformToElement(rootSVGElement, path1).f should be close to -18.9. Was -17.677916691624347.
PASS getTransformToElement(rootSVGElement, path2).a is -0.7
PASS getTransformToElement(rootSVGElement, path2).b is 0.7
PASS getTransformToElement(rootSVGElement, path2).c is -0.7
PASS getTransformToElement(rootSVGElement, path2).d is -0.7
FAIL getTransformToElement(rootSVGElement, path2).e should be close to 442.7. Was 441.941810617502.
FAIL getTransformToElement(rootSVGElement, path2).f should be close to -156.9. Was -159.09852261339003.
PASS getTransformToElement(rootSVGElement, path1).a is 0.7
PASS getTransformToElement(rootSVGElement, path1).b is -0.7
PASS getTransformToElement(rootSVGElement, path1).c is 0.7
PASS getTransformToElement(rootSVGElement, path1).d is 0.7
PASS getTransformToElement(rootSVGElement, path1).e is -265.1
FAIL getTransformToElement(rootSVGElement, path1).f should be close to -18.9. Was -17.677916691624347.
PASS getTransformToElement(rootSVGElement, path2).a is -0.7
PASS getTransformToElement(rootSVGElement, path2).b is 0.7
PASS getTransformToElement(rootSVGElement, path2).c is -0.7
PASS getTransformToElement(rootSVGElement, path2).d is -0.7
FAIL getTransformToElement(rootSVGElement, path2).e should be close to 442.7. Was 441.941810617502.
FAIL getTransformToElement(rootSVGElement, path2).f should be close to -156.9. Was -159.09852261339003.
PASS successfullyParsed is true

TEST COMPLETE
